+++ beacon-chain/core/balances/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
     deps = [
         "//beacon-chain/core/epoch:go_default_library",
         "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/core/validators:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/params:go_default_library",
         "//shared/sliceutil:go_default_library",
