+++ beacon-chain/core/balances/rewards_penalties.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/epoch"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/validators"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/sliceutil"
@@ -300,7 +299,7 @@ func AttestationInclusion(
 		if err != nil {
 			return nil, fmt.Errorf("could not get inclusion slot: %v", err)
 		}
-		proposerIndex, err := validators.BeaconProposerIdx(state, slot)
+		proposerIndex, err := helpers.BeaconProposerIndex(state, slot)
 		if err != nil {
 			return nil, fmt.Errorf("could not get propoer index: %v", err)
 		}
