+++ beacon-chain/core/validators/validator.go
@@ -30,45 +30,8 @@ func InitialValidatorRegistry() []*pb.Validator {
 	return validators
 }
 
-// ActiveValidators returns the active validator records in a list.
-//
-// Spec pseudocode definition:
-//   [state.validator_registry[i] for i in get_active_validator_indices(state.validator_registry)]
-func ActiveValidators(state *pb.BeaconState, validatorIndices []uint32) []*pb.Validator {
-	activeValidators := make([]*pb.Validator, 0, len(validatorIndices))
-	for _, validatorIndex := range validatorIndices {
-		activeValidators = append(activeValidators, state.ValidatorRegistry[validatorIndex])
-	}
-	return activeValidators
-}
-
-// BeaconProposerIdx returns the index of the proposer of the block at a
-// given slot.
-//
-// Spec pseudocode definition:
-//  def get_beacon_proposer_index(state: BeaconState,slot: int) -> int:
-//    """
-//    Returns the beacon proposer index for the ``slot``.
-//    """
-//    first_committee, _ = get_crosslink_committees_at_slot(state, slot)[0]
-//    return first_committee[slot % len(first_committee)]
-func BeaconProposerIdx(state *pb.BeaconState, slot uint64) (uint64, error) {
-	committeeArray, err := helpers.CrosslinkCommitteesAtSlot(state, slot, false)
-	if err != nil {
-		return 0, err
-	}
-	firstCommittee := committeeArray[0].Committee
-
-	if len(firstCommittee) == 0 {
-		return 0, fmt.Errorf("empty first committee at slot %d", slot)
-	}
-
-	return firstCommittee[slot%uint64(len(firstCommittee))], nil
-}
-
 // ValidatorIdx returns the idx of the validator given an input public key.
 func ValidatorIdx(pubKey []byte, validators []*pb.Validator) (uint64, error) {
-
 	for idx := range validators {
 		if bytes.Equal(validators[idx].Pubkey, pubKey) {
 			return uint64(idx), nil
@@ -76,35 +39,6 @@ func ValidatorIdx(pubKey []byte, validators []*pb.Validator) (uint64, error) {
 	}
 
 	return 0, fmt.Errorf("can't find validator index for public key %#x", pubKey)
-}
-
-// TotalEffectiveBalance returns the total deposited amount at stake in Gwei
-// of all active validators.
-//
-// Spec pseudocode definition:
-//   sum([get_effective_balance(state, i) for i in active_validator_indices])
-func TotalEffectiveBalance(state *pb.BeaconState, validatorIndices []uint64) uint64 {
-	var totalDeposit uint64
-
-	for _, idx := range validatorIndices {
-		totalDeposit += helpers.EffectiveBalance(state, idx)
-	}
-	return totalDeposit
-}
-
-// Attesters returns the validator records using validator indices.
-//
-// Spec pseudocode definition:
-//   Let current_epoch_boundary_attesters = [state.validator_registry[i]
-//   for indices in current_epoch_boundary_attester_indices for i in indices].
-func Attesters(state *pb.BeaconState, attesterIndices []uint64) []*pb.Validator {
-
-	var boundaryAttesters []*pb.Validator
-	for _, attesterIdx := range attesterIndices {
-		boundaryAttesters = append(boundaryAttesters, state.ValidatorRegistry[attesterIdx])
-	}
-
-	return boundaryAttesters
 }
 
 // ValidatorIndices returns all the validator indices from the input attestations
@@ -135,15 +69,15 @@ func ValidatorIndices(
 	return attesterIndicesIntersection, nil
 }
 
-// AttestingValidatorIndices returns the shard committee validator indices
-// if the validator shard committee matches the input attestations.
+// AttestingValidatorIndices returns the crosslink committee validator indices
+// if the validators from crosslink committee is part of the input attestations.
 //
 // Spec pseudocode definition:
 // Let attesting_validator_indices(crosslink_committee, shard_block_root)
-// be the union of the validator index sets given by
-// [get_attestation_participants(state, a.data, a.participation_bitfield)
-// for a in current_epoch_attestations + previous_epoch_attestations
-// if a.shard == shard_committee.shard and a.shard_block_root == shard_block_root]
+// 	be the union of the validator index sets given by
+// 	[get_attestation_participants(state, a.data, a.participation_bitfield)
+// 	for a in current_epoch_attestations + previous_epoch_attestations
+// 		if a.shard == shard_committee.shard and a.shard_block_root == shard_block_root]
 func AttestingValidatorIndices(
 	state *pb.BeaconState,
 	shard uint64,
@@ -166,32 +100,6 @@ func AttestingValidatorIndices(
 		}
 	}
 	return validatorIndicesCommittees, nil
-}
-
-// AttestingBalance returns the combined balances from the input validator
-// records.
-//
-// Spec pseudocode definition:
-//   Let current_epoch_boundary_attesting_balance =
-//   sum([get_effective_balance(state, i) for i in current_epoch_boundary_attester_indices])
-func AttestingBalance(state *pb.BeaconState, boundaryAttesterIndices []uint64) uint64 {
-
-	var boundaryAttestingBalance uint64
-	for _, idx := range boundaryAttesterIndices {
-		boundaryAttestingBalance += helpers.EffectiveBalance(state, idx)
-	}
-
-	return boundaryAttestingBalance
-}
-
-// AllValidatorsIndices returns all validator indices from 0 to
-// the last validator.
-func AllValidatorsIndices(state *pb.BeaconState) []uint64 {
-	validatorIndices := make([]uint64, len(state.ValidatorRegistry))
-	for i := 0; i < len(validatorIndices); i++ {
-		validatorIndices[i] = uint64(i)
-	}
-	return validatorIndices
 }
 
 // ProcessDeposit mutates a corresponding index in the beacon state for
@@ -298,7 +206,7 @@ func InitiateValidatorExit(state *pb.BeaconState, idx uint64) *pb.BeaconState {
 func ExitValidator(state *pb.BeaconState, idx uint64) (*pb.BeaconState, error) {
 	validator := state.ValidatorRegistry[idx]
 
-	exitEpoch := helpers.EntryExitEffectEpoch(helpers.CurrentEpoch(state))
+	exitEpoch := entryExitEffectEpoch(helpers.CurrentEpoch(state))
 	if validator.ExitEpoch <= exitEpoch {
 		return nil, fmt.Errorf("validator %d could not exit until epoch %d",
 			idx, exitEpoch)
@@ -335,7 +243,7 @@ func PenalizeValidator(state *pb.BeaconState, idx uint64) (*pb.BeaconState, erro
 	penalizedDuration := helpers.CurrentEpoch(state) % params.BeaconConfig().LatestPenalizedExitLength
 	state.LatestPenalizedBalances[penalizedDuration] += helpers.EffectiveBalance(state, idx)
 
-	whistleblowerIdx, err := BeaconProposerIdx(state, state.Slot)
+	whistleblowerIdx, err := helpers.BeaconProposerIndex(state, state.Slot)
 	if err != nil {
 		return nil, fmt.Errorf("could not get proposer idx: %v", err)
 	}
@@ -347,23 +255,6 @@ func PenalizeValidator(state *pb.BeaconState, idx uint64) (*pb.BeaconState, erro
 
 	state.ValidatorRegistry[idx].PenalizedEpoch = helpers.CurrentEpoch(state)
 	return state, nil
-}
-
-// PrepareValidatorForWithdrawal sets validator's status flag to
-// WITHDRAWABLE.
-//
-// Spec pseudocode definition:
-// def prepare_validator_for_withdrawal(state: BeaconState, index: ValidatorIndex) -> None:
-//    """
-//    Set the validator with the given ``index`` with ``WITHDRAWABLE`` flag.
-//    Note that this function mutates ``state``.
-//    """
-//    validator = state.validator_registry[index]
-//    validator.status_flags |= WITHDRAWABLE
-func PrepareValidatorForWithdrawal(state *pb.BeaconState, idx uint64) *pb.BeaconState {
-	state.ValidatorRegistry[idx].StatusFlags |=
-		pb.Validator_WITHDRAWABLE
-	return state
 }
 
 // UpdateRegistry rotates validators in and out of active pool.
@@ -417,7 +308,7 @@ func UpdateRegistry(state *pb.BeaconState) (*pb.BeaconState, error) {
 	activeValidatorIndices := helpers.ActiveValidatorIndices(
 		state.ValidatorRegistry, currentEpoch)
 
-	totalBalance := TotalEffectiveBalance(state, activeValidatorIndices)
+	totalBalance := helpers.TotalBalance(state, activeValidatorIndices)
 
 	// The maximum balance churn in Gwei (for deposits and exits separately).
 	maxBalChurn := maxBalanceChurn(totalBalance)
@@ -504,7 +395,7 @@ func ProcessPenaltiesAndExits(state *pb.BeaconState) *pb.BeaconState {
 	currentEpoch := helpers.CurrentEpoch(state)
 	activeValidatorIndices := helpers.ActiveValidatorIndices(
 		state.ValidatorRegistry, currentEpoch)
-	totalBalance := TotalEffectiveBalance(state, activeValidatorIndices)
+	totalBalance := helpers.TotalBalance(state, activeValidatorIndices)
 
 	for idx, validator := range state.ValidatorRegistry {
 		penalized := validator.PenalizedEpoch +
@@ -525,7 +416,7 @@ func ProcessPenaltiesAndExits(state *pb.BeaconState) *pb.BeaconState {
 			state.ValidatorBalances[idx] -= penalty
 		}
 	}
-	allIndices := AllValidatorsIndices(state)
+	allIndices := allValidatorsIndices(state)
 	var eligibleIndices []uint64
 	for _, idx := range allIndices {
 		if eligibleToExit(state, idx) {
@@ -534,13 +425,23 @@ func ProcessPenaltiesAndExits(state *pb.BeaconState) *pb.BeaconState {
 	}
 	var withdrawnSoFar uint64
 	for _, idx := range eligibleIndices {
-		state = PrepareValidatorForWithdrawal(state, idx)
+		state = prepareValidatorForWithdrawal(state, idx)
 		withdrawnSoFar++
 		if withdrawnSoFar >= params.BeaconConfig().MaxWithdrawalsPerEpoch {
 			break
 		}
 	}
 	return state
+}
+
+// allValidatorsIndices returns all validator indices from 0 to
+// the last validator.
+func allValidatorsIndices(state *pb.BeaconState) []uint64 {
+	validatorIndices := make([]uint64, len(state.ValidatorRegistry))
+	for i := 0; i < len(validatorIndices); i++ {
+		validatorIndices[i] = uint64(i)
+	}
+	return validatorIndices
 }
 
 // maxBalanceChurn returns the maximum balance churn in Gwei,
@@ -578,4 +479,35 @@ func eligibleToExit(state *pb.BeaconState, idx uint64) bool {
 		return currentEpoch >= validator.PenalizedEpoch+penalizedWithdrawalEpochs
 	}
 	return currentEpoch >= validator.ExitEpoch+params.BeaconConfig().MinValidatorWithdrawalEpochs
+}
+
+// prepareValidatorForWithdrawal sets validator's status flag to
+// WITHDRAWABLE.
+//
+// Spec pseudocode definition:
+// def prepare_validator_for_withdrawal(state: BeaconState, index: ValidatorIndex) -> None:
+//    """
+//    Set the validator with the given ``index`` with ``WITHDRAWABLE`` flag.
+//    Note that this function mutates ``state``.
+//    """
+//    validator = state.validator_registry[index]
+//    validator.status_flags |= WITHDRAWABLE
+func prepareValidatorForWithdrawal(state *pb.BeaconState, idx uint64) *pb.BeaconState {
+	state.ValidatorRegistry[idx].StatusFlags |=
+		pb.Validator_WITHDRAWABLE
+	return state
+}
+
+// entryExitEffectEpoch takes in epoch number and returns when
+// the validator is eligible for activation and exit.
+//
+// Spec pseudocode definition:
+// def get_entry_exit_effect_epoch(epoch: EpochNumber) -> EpochNumber:
+//    """
+//    An entry or exit triggered in the ``epoch`` given by the input takes effect at
+//    the epoch given by the output.
+//    """
+//    return epoch + 1 + ENTRY_EXIT_DELAY
+func entryExitEffectEpoch(epoch uint64) uint64 {
+	return epoch + 1 + params.BeaconConfig().EntryExitDelay
 }
