+++ beacon-chain/core/validators/validator_test.go
@@ -76,102 +76,6 @@ func TestValidatorIdx(t *testing.T) {
 	}
 }
 
-func TestEffectiveBalance(t *testing.T) {
-	defaultBalance := params.BeaconConfig().MaxDepositAmount
-
-	tests := []struct {
-		a uint64
-		b uint64
-	}{
-		{a: 0, b: 0},
-		{a: defaultBalance - 1, b: defaultBalance - 1},
-		{a: defaultBalance, b: defaultBalance},
-		{a: defaultBalance + 1, b: defaultBalance},
-		{a: defaultBalance * 100, b: defaultBalance},
-	}
-	for _, test := range tests {
-		state := &pb.BeaconState{ValidatorBalances: []uint64{test.a}}
-		if helpers.EffectiveBalance(state, 0) != test.b {
-			t.Errorf("EffectiveBalance(%d) = %d, want = %d", test.a, helpers.EffectiveBalance(state, 0), test.b)
-		}
-	}
-}
-
-func TestTotalEffectiveBalance(t *testing.T) {
-	state := &pb.BeaconState{ValidatorBalances: []uint64{
-		27 * 1e9, 28 * 1e9, 32 * 1e9, 40 * 1e9,
-	}}
-
-	// 27 + 28 + 32 + 32 = 119
-	if TotalEffectiveBalance(state, []uint64{0, 1, 2, 3}) != 119*1e9 {
-		t.Errorf("Incorrect TotalEffectiveBalance. Wanted: 119, got: %d",
-			TotalEffectiveBalance(state, []uint64{0, 1, 2, 3})/1e9)
-	}
-}
-
-func TestGetActiveValidator(t *testing.T) {
-	inputValidators := []*pb.Validator{
-		{Pubkey: []byte("A")},
-		{Pubkey: []byte("B")},
-		{Pubkey: []byte("C")},
-		{Pubkey: []byte("D")},
-		{Pubkey: []byte("E")},
-	}
-
-	outputValidators := []*pb.Validator{
-		{Pubkey: []byte("B")},
-		{Pubkey: []byte("D")},
-	}
-
-	state := &pb.BeaconState{
-		ValidatorRegistry: inputValidators,
-	}
-
-	validators := ActiveValidators(state, []uint32{1, 3})
-
-	if !reflect.DeepEqual(outputValidators, validators) {
-		t.Errorf("Active validators don't match. Wanted: %v, Got: %v", outputValidators, validators)
-	}
-}
-
-func TestBoundaryAttestingBalance(t *testing.T) {
-	state := &pb.BeaconState{ValidatorBalances: []uint64{
-		25 * 1e9, 26 * 1e9, 32 * 1e9, 33 * 1e9, 100 * 1e9,
-	}}
-
-	attestedBalances := AttestingBalance(state, []uint64{0, 1, 2, 3, 4})
-
-	// 25 + 26 + 32 + 32 + 32 = 147
-	if attestedBalances != 147*1e9 {
-		t.Errorf("Incorrect attested balances. Wanted: %f, got: %d", 147*1e9, attestedBalances)
-	}
-}
-
-func TestBoundaryAttesters(t *testing.T) {
-	var validators []*pb.Validator
-
-	for i := 0; i < 100; i++ {
-		validators = append(validators, &pb.Validator{Pubkey: []byte{byte(i)}})
-	}
-
-	state := &pb.BeaconState{ValidatorRegistry: validators}
-
-	boundaryAttesters := Attesters(state, []uint64{5, 2, 87, 42, 99, 0})
-
-	expectedBoundaryAttesters := []*pb.Validator{
-		{Pubkey: []byte{byte(5)}},
-		{Pubkey: []byte{byte(2)}},
-		{Pubkey: []byte{byte(87)}},
-		{Pubkey: []byte{byte(42)}},
-		{Pubkey: []byte{byte(99)}},
-		{Pubkey: []byte{byte(0)}},
-	}
-
-	if !reflect.DeepEqual(expectedBoundaryAttesters, boundaryAttesters) {
-		t.Errorf("Incorrect boundary attesters. Wanted: %v, got: %v", expectedBoundaryAttesters, boundaryAttesters)
-	}
-}
-
 func TestBoundaryAttesterIndices(t *testing.T) {
 	if params.BeaconConfig().EpochLength != 64 {
 		t.Errorf("EpochLength should be 64 for these tests to pass")
@@ -200,72 +104,6 @@ func TestBoundaryAttesterIndices(t *testing.T) {
 	if !reflect.DeepEqual(attesterIndices, []uint64{109, 97}) {
 		t.Errorf("Incorrect boundary attester indices. Wanted: %v, got: %v",
 			[]uint64{109, 97}, attesterIndices)
-	}
-}
-
-func TestBeaconProposerIdx(t *testing.T) {
-	if params.BeaconConfig().EpochLength != 64 {
-		t.Errorf("EpochLength should be 64 for these tests to pass")
-	}
-
-	validators := make([]*pb.Validator, params.BeaconConfig().DepositsForChainStart)
-	for i := 0; i < len(validators); i++ {
-		validators[i] = &pb.Validator{
-			ExitEpoch: params.BeaconConfig().FarFutureEpoch,
-		}
-	}
-
-	state := &pb.BeaconState{
-		ValidatorRegistry: validators,
-	}
-
-	tests := []struct {
-		slot  uint64
-		index uint64
-	}{
-		{
-			slot:  1,
-			index: 511,
-		},
-		{
-			slot:  10,
-			index: 2807,
-		},
-		{
-			slot:  19,
-			index: 5122,
-		},
-		{
-			slot:  30,
-			index: 7947,
-		},
-		{
-			slot:  39,
-			index: 10262,
-		},
-	}
-
-	for _, tt := range tests {
-		result, err := BeaconProposerIdx(state, tt.slot)
-		if err != nil {
-			t.Errorf("Failed to get shard and committees at slot: %v", err)
-		}
-
-		if result != tt.index {
-			t.Errorf(
-				"Result index was an unexpected value. Wanted %d, got %d",
-				tt.index,
-				result,
-			)
-		}
-	}
-}
-
-func TestBeaconProposerIdx_returnsErrorWithEmptyCommittee(t *testing.T) {
-	_, err := BeaconProposerIdx(&pb.BeaconState{}, 0)
-	expected := "empty first committee at slot 0"
-	if err.Error() != expected {
-		t.Errorf("Unexpected error. got=%v want=%s", err, expected)
 	}
 }
 
@@ -357,9 +195,9 @@ func TestAllValidatorIndices(t *testing.T) {
 	}
 	for _, tt := range tests {
 		state := &pb.BeaconState{ValidatorRegistry: tt.registries}
-		if !reflect.DeepEqual(AllValidatorsIndices(state), tt.indices) {
+		if !reflect.DeepEqual(allValidatorsIndices(state), tt.indices) {
 			t.Errorf("AllValidatorsIndices(%v) = %v, wanted:%v",
-				tt.registries, AllValidatorsIndices(state), tt.indices)
+				tt.registries, allValidatorsIndices(state), tt.indices)
 		}
 	}
 }
